[feat]: Implementado o método cadastrar gerando uma senha aleatória e enviando por email
